feat: 선택 페이지 추가
